+++ app/controllers/entities_controller.rb
@@ -37,7 +37,7 @@ class EntitiesController < ApplicationController
       @group = SystemGroup.find_or_create_by(:name => 'invalid')
       @entities = @group.entities.allowed(current_user, :delete).paginate :page => params[:page], :per_page => 30
     else
-      render_denied_page
+      render_403
     end
   end
   
@@ -48,7 +48,7 @@ class EntitiesController < ApplicationController
         per_page: 30
       )
     else
-      render_denied_page
+      render_403
     end
   end
 
@@ -191,7 +191,7 @@ class EntitiesController < ApplicationController
       @entity.no_name_statement = 'enter_name'
       @entity.medium = Medium.new if @entity.kind_id == Kind.medium_kind.id
     else
-      render_denied_page
+      render_403
     end
   end
   
@@ -206,7 +206,7 @@ class EntitiesController < ApplicationController
     if authorized? :edit, @entity.collection
       render :action => 'edit'  
     else
-      render_denied_page
+      render_403
     end
   end
 
@@ -258,7 +258,7 @@ class EntitiesController < ApplicationController
         end
       end
     else
-      render_denied_page
+      render_403
     end
   end
 
@@ -287,7 +287,7 @@ class EntitiesController < ApplicationController
         render action: "edit"
       end
     else
-      render_denied_page
+      render_403
     end
   rescue ActiveRecord::StaleObjectError
     flash[:error] = I18n.t('activerecord.errors.messages.stale_entity_update')
@@ -300,7 +300,7 @@ class EntitiesController < ApplicationController
       @entity.destroy
       redirect_to back_save
     else
-      render_denied_page
+      render_403
     end
   end