+++ app/controllers/media_controller.rb
@@ -6,7 +6,7 @@ class MediaController < ApplicationController
     if authorized?(:view, @medium.entity.collection)
       render :layout => 'wide'
     else
-      render_denied_page
+      render_403
     end
   end
 
@@ -30,7 +30,7 @@ class MediaController < ApplicationController
         :disposition => 'inline',
         :filename => @medium.download_filename(style)
     else
-      render_denied_page
+      render_403
     end
   end
 
@@ -53,7 +53,7 @@ class MediaController < ApplicationController
         :disposition => 'attachment',
         :filename => @medium.download_filename(style)
     else
-      render_denied_page
+      render_403
     end
   end
 
@@ -68,7 +68,7 @@ class MediaController < ApplicationController
 
       redirect_to web_path(:anchor => entity_path(@medium.entity))
     else
-      render_denied_page
+      render_403
     end
   end