+++ app/controllers/tools_controller.rb
@@ -88,6 +88,8 @@ class ToolsController < ApplicationController
       format.html do 
         if request.referer && request.referer.match(/\/blaze/)
           redirect_to back_save
+        elsif request.referer && request.referer.match(/\/authority_group/)
+          redirect_to :back
         else
           redirect_to :controller => 'tools', :action => 'clipboard'
         end
@@ -189,7 +191,7 @@ class ToolsController < ApplicationController
         flash[:notice] = I18n.t('messages.entities_moved_to_collection', :o => collection.name)
         redirect_to clipboard_path
       else
-        render_denied_page
+        render_403
       end
     end
   
@@ -210,7 +212,7 @@ class ToolsController < ApplicationController
           redirect_to back_save
         end
       else
-        render_denied_page
+        render_403
       end
     end
 
@@ -226,7 +228,7 @@ class ToolsController < ApplicationController
           redirect_to back_save
         end
       else
-        render_denied_page
+        render_403
       end
     end
 
@@ -264,7 +266,7 @@ class ToolsController < ApplicationController
       
       if @entities.blank?
         flash[:error] = I18n.t("errors.merge_access_denied_on_entities")
-        render_denied_page
+        render_403
       elsif @entities.collect{|e| e.kind.id}.uniq.size != 1
         flash[:error] = I18n.t("errors.only_same_kind")
         redirect_to :controller => 'tools', :action => 'clipboard', :entity_ids => params[:entity_ids]
@@ -326,7 +328,7 @@ class ToolsController < ApplicationController
           redirect_to :action => 'clipboard'
         end
       else
-        render_denied_page
+        render_403
       end
     end